+++ influxql/iterator.go
@@ -232,6 +232,10 @@ type AuxIterator interface {
 
 	// Start starts writing to the created iterators.
 	Start()
+
+	// Backgrounds the iterator so that, when start is called, it will
+	// continuously read from the iterator.
+	Background()
 }
 
 // NewAuxIterator returns a new instance of AuxIterator.
@@ -312,19 +316,19 @@ func (a auxIteratorFields) iterator(name string) Iterator {
 		// Create channel iterator by data type.
 		switch f.typ {
 		case Float:
-			itr := &floatChanIterator{c: make(chan *FloatPoint, 1)}
+			itr := &floatChanIterator{cond: sync.NewCond(&sync.Mutex{})}
 			f.append(itr)
 			return itr
 		case Integer:
-			itr := &integerChanIterator{c: make(chan *IntegerPoint, 1)}
+			itr := &integerChanIterator{cond: sync.NewCond(&sync.Mutex{})}
 			f.append(itr)
 			return itr
 		case String:
-			itr := &stringChanIterator{c: make(chan *StringPoint, 1)}
+			itr := &stringChanIterator{cond: sync.NewCond(&sync.Mutex{})}
 			f.append(itr)
 			return itr
 		case Boolean:
-			itr := &booleanChanIterator{c: make(chan *BooleanPoint, 1)}
+			itr := &booleanChanIterator{cond: sync.NewCond(&sync.Mutex{})}
 			f.append(itr)
 			return itr
 		default:
@@ -336,7 +340,7 @@ func (a auxIteratorFields) iterator(name string) Iterator {
 }
 
 // send sends a point to all field iterators.
-func (a auxIteratorFields) send(p Point) {
+func (a auxIteratorFields) send(p Point) (ok bool) {
 	values := p.aux()
 	for i, f := range a {
 		v := values[i]
@@ -349,40 +353,19 @@ func (a auxIteratorFields) send(p Point) {
 		for _, itr := range f.itrs {
 			switch itr := itr.(type) {
 			case *floatChanIterator:
-				switch v := v.(type) {
-				case float64:
-					itr.c <- &FloatPoint{Name: p.name(), Tags: tags, Time: p.time(), Value: v}
-				case int64:
-					itr.c <- &FloatPoint{Name: p.name(), Tags: tags, Time: p.time(), Value: float64(v)}
-				default:
-					itr.c <- &FloatPoint{Name: p.name(), Tags: tags, Time: p.time(), Nil: true}
-				}
+				ok = itr.setBuf(p.name(), tags, p.time(), v) || ok
 			case *integerChanIterator:
-				switch v := v.(type) {
-				case int64:
-					itr.c <- &IntegerPoint{Name: p.name(), Tags: tags, Time: p.time(), Value: v}
-				default:
-					itr.c <- &IntegerPoint{Name: p.name(), Tags: tags, Time: p.time(), Nil: true}
-				}
+				ok = itr.setBuf(p.name(), tags, p.time(), v) || ok
 			case *stringChanIterator:
-				switch v := v.(type) {
-				case string:
-					itr.c <- &StringPoint{Name: p.name(), Tags: tags, Time: p.time(), Value: v}
-				default:
-					itr.c <- &StringPoint{Name: p.name(), Tags: tags, Time: p.time(), Nil: true}
-				}
+				ok = itr.setBuf(p.name(), tags, p.time(), v) || ok
 			case *booleanChanIterator:
-				switch v := v.(type) {
-				case bool:
-					itr.c <- &BooleanPoint{Name: p.name(), Tags: tags, Time: p.time(), Value: v}
-				default:
-					itr.c <- &BooleanPoint{Name: p.name(), Tags: tags, Time: p.time(), Nil: true}
-				}
+				ok = itr.setBuf(p.name(), tags, p.time(), v) || ok
 			default:
 				panic(fmt.Sprintf("invalid aux itr type: %T", itr))
 			}
 		}
 	}
+	return ok
 }
 
 // drainIterator reads all points from an iterator.
@@ -444,6 +427,9 @@ type IteratorCreator interface {
 
 	// Returns the series keys that will be returned by this iterator.
 	SeriesKeys(opt IteratorOptions) (SeriesList, error)
+
+	// Expands regex sources to all matching sources.
+	ExpandSources(sources Sources) (Sources, error)
 }
 
 // IteratorCreators represents a list of iterator creators.
@@ -542,6 +528,42 @@ func (a IteratorCreators) SeriesKeys(opt IteratorOptions) (SeriesList, error) {
 	}
 	sort.Sort(SeriesList(seriesList))
 	return SeriesList(seriesList), nil
+}
+
+// ExpandSources expands sources across all iterator creators and returns a unique result.
+func (a IteratorCreators) ExpandSources(sources Sources) (Sources, error) {
+	m := make(map[string]Source)
+
+	for _, ic := range a {
+		expanded, err := ic.ExpandSources(sources)
+		if err != nil {
+			return nil, err
+		}
+
+		for _, src := range expanded {
+			switch src := src.(type) {
+			case *Measurement:
+				m[src.String()] = src
+			default:
+				return nil, fmt.Errorf("IteratorCreators.ExpandSources: unsupported source type: %T", src)
+			}
+		}
+	}
+
+	// Convert set to sorted slice.
+	names := make([]string, 0, len(m))
+	for name := range m {
+		names = append(names, name)
+	}
+	sort.Strings(names)
+
+	// Convert set to a list of Sources.
+	sorted := make(Sources, 0, len(m))
+	for _, name := range names {
+		sorted = append(sorted, m[name])
+	}
+
+	return sorted, nil
 }
 
 // IteratorOptions is an object passed to CreateIterator to specify creation options.