+++ services/httpd/handler.go
@@ -8,7 +8,6 @@ import (
 	"expvar"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"net/http/pprof"
@@ -20,7 +19,6 @@ import (
 	"github.com/bmizerany/pat"
 	"github.com/influxdata/influxdb"
 	"github.com/influxdata/influxdb/client"
-	"github.com/influxdata/influxdb/cluster"
 	"github.com/influxdata/influxdb/influxql"
 	"github.com/influxdata/influxdb/models"
 	"github.com/influxdata/influxdb/services/continuous_querier"
@@ -61,14 +59,16 @@ type Handler struct {
 		Database(name string) (*meta.DatabaseInfo, error)
 		Authenticate(username, password string) (ui *meta.UserInfo, err error)
 		Users() []meta.UserInfo
-		Ping(checkAllMetaServers bool) error
 	}
 
-	QueryAuthorizer *meta.QueryAuthorizer
-	QueryExecutor   influxql.QueryExecutor
+	QueryAuthorizer interface {
+		AuthorizeQuery(u *meta.UserInfo, query *influxql.Query, database string) error
+	}
+
+	QueryExecutor influxql.QueryExecutor
 
 	PointsWriter interface {
-		WritePoints(p *cluster.WritePointsRequest) error
+		WritePoints(database, retentionPolicy string, consistencyLevel models.ConsistencyLevel, points []models.Point) error
 	}
 
 	ContinuousQuerier continuous_querier.ContinuousQuerier
@@ -389,7 +389,17 @@ func (h *Handler) serveWrite(w http.ResponseWriter, r *http.Request, user *meta.
 		body = b
 	}
 
-	b, err := ioutil.ReadAll(body)
+	var bs []byte
+	if clStr := r.Header.Get("Content-Length"); clStr != "" {
+		if length, err := strconv.Atoi(clStr); err == nil {
+			// This will just be an initial hint for the gzip reader, as the
+			// bytes.Buffer will grow as needed when ReadFrom is called
+			bs = make([]byte, 0, length)
+		}
+	}
+	buf := bytes.NewBuffer(bs)
+
+	_, err := buf.ReadFrom(body)
 	if err != nil {
 		if h.WriteTrace {
 			h.Logger.Print("write handler unable to read bytes from request body")
@@ -397,16 +407,17 @@ func (h *Handler) serveWrite(w http.ResponseWriter, r *http.Request, user *meta.
 		resultError(w, influxql.Result{Err: err}, http.StatusBadRequest)
 		return
 	}
-	h.statMap.Add(statWriteRequestBytesReceived, int64(len(b)))
+	h.statMap.Add(statWriteRequestBytesReceived, int64(buf.Len()))
+
 	if h.WriteTrace {
-		h.Logger.Printf("write body received by handler: %s", string(b))
+		h.Logger.Printf("write body received by handler: %s", buf.Bytes())
 	}
 
 	if r.Header.Get("Content-Type") == "application/json" {
-		h.serveWriteJSON(w, r, b, user)
+		h.serveWriteJSON(w, r, buf.Bytes(), user)
 		return
 	}
-	h.serveWriteLine(w, r, b, user)
+	h.serveWriteLine(w, r, buf.Bytes(), user)
 }
 
 // serveWriteJSON receives incoming series data in JSON and writes it to the database.
@@ -460,12 +471,7 @@ func (h *Handler) serveWriteJSON(w http.ResponseWriter, r *http.Request, body []
 	}
 
 	// Convert the json batch struct to a points writer struct
-	if err := h.PointsWriter.WritePoints(&cluster.WritePointsRequest{
-		Database:         bp.Database,
-		RetentionPolicy:  bp.RetentionPolicy,
-		ConsistencyLevel: cluster.ConsistencyLevelOne,
-		Points:           points,
-	}); err != nil {
+	if err := h.PointsWriter.WritePoints(bp.Database, bp.RetentionPolicy, models.ConsistencyLevelAny, points); err != nil {
 		h.statMap.Add(statPointsWrittenFail, int64(len(points)))
 		if influxdb.IsClientError(err) {
 			resultError(w, influxql.Result{Err: err}, http.StatusBadRequest)
@@ -540,26 +546,8 @@ func (h *Handler) serveWriteLine(w http.ResponseWriter, r *http.Request, body []
 		return
 	}
 
-	// Determine required consistency level.
-	consistency := cluster.ConsistencyLevelOne
-	switch r.Form.Get("consistency") {
-	case "all":
-		consistency = cluster.ConsistencyLevelAll
-	case "any":
-		consistency = cluster.ConsistencyLevelAny
-	case "one":
-		consistency = cluster.ConsistencyLevelOne
-	case "quorum":
-		consistency = cluster.ConsistencyLevelQuorum
-	}
-
 	// Write points.
-	if err := h.PointsWriter.WritePoints(&cluster.WritePointsRequest{
-		Database:         database,
-		RetentionPolicy:  r.FormValue("rp"),
-		ConsistencyLevel: consistency,
-		Points:           points,
-	}); influxdb.IsClientError(err) {
+	if err := h.PointsWriter.WritePoints(database, r.FormValue("rp"), models.ConsistencyLevelAny, points); influxdb.IsClientError(err) {
 		h.statMap.Add(statPointsWrittenFail, int64(len(points)))
 		resultError(w, influxql.Result{Err: err}, http.StatusBadRequest)
 		return
@@ -591,15 +579,10 @@ func (h *Handler) servePing(w http.ResponseWriter, r *http.Request) {
 	w.WriteHeader(http.StatusNoContent)
 }
 
-// serveStatus returns a simple response to let the client know the whole cluster is running.
+// serveStatus has been depricated
 func (h *Handler) serveStatus(w http.ResponseWriter, r *http.Request) {
+	h.Logger.Printf("WARNING: /status has been depricated.  Use /ping instead.")
 	h.statMap.Add(statStatusRequest, 1)
-
-	if err := h.MetaClient.Ping(false); err != nil {
-		w.WriteHeader(http.StatusServiceUnavailable)
-		return
-	}
-
 	w.WriteHeader(http.StatusNoContent)
 }